tr2/decomp: move skidoo code to the vehicles module
