+++ src/tr2/game/objects/vehicles/skidoo_common.c
@@ -1,6 +1,5 @@
-#include "decomp/skidoo.h"
+#include "game/objects/vehicles/skidoo_common.h"
 
-#include "decomp/decomp.h"
 #include "game/creature.h"
 #include "game/effects.h"
 #include "game/objects/common.h"
