+++ src/tr2/game/objects/vehicles/skidoo_common.h
